change data type

+++ internal/data/api/review_depth.go
@@ -8,11 +8,11 @@ import (
 )
 
 type MRReviewDepth struct {
-	Week         string  `json:"week"`
-	Average      float64 `json:"average"`
-	Median       float64 `json:"median"`
-	Percentile75 float64 `json:"percentile75"`
-	Percentile95 float64 `json:"percentile95"`
+	Week         string `json:"week"`
+	Average      int    `json:"average"`
+	Median       int    `json:"median"`
+	Percentile75 int    `json:"percentile75"`
+	Percentile95 int    `json:"percentile95"`
 }
 
 /*
